Added support for code loaded by AutoLoader and SelfLoader.
Core modules AutoLoader and SelfLoader support delayed compilation by
allowing Perl code that follows __END__ and __DATA__ tokens, respectively.
When we detect that one of these modules is used in the file, we continue
processing even after we see the corresponding token that would usually
terminate parsing of the file.